fixed do_sample_* methods
