+++ sympy/stats/sampling/sample_scipy.py
@@ -24,7 +24,7 @@ def do_sample_scipy(dist, size, seed):
 
 # CRV
 
-@do_sample_scipy.register
+@do_sample_scipy.register(SingleContinuousDistribution)
 def _(dist: SingleContinuousDistribution, size, seed):
     # if we don't need to make a handmade pdf, we won't
     import scipy.stats
@@ -41,66 +41,66 @@ def _(dist: SingleContinuousDistribution, size, seed):
     return scipy_rv.rvs(size=size, random_state=seed)
 
 
-@do_sample_scipy.register
+@do_sample_scipy.register(ChiSquaredDistribution)
 def _(dist: ChiSquaredDistribution, size, seed):
     # same parametrisation
     return scipy.stats.chi2.rvs(df=float(dist.k), size=size, random_state=seed)
 
 
-@do_sample_scipy.register
+@do_sample_scipy.register(ExponentialDistribution)
 def _(dist: ExponentialDistribution, size, seed):
     # https://docs.scipy.org/doc/scipy/reference/generated/scipy.stats.expon.html#scipy.stats.expon
     return scipy.stats.expon.rvs(scale=1 / float(dist.rate), size=size, random_state=seed)
 
 
-@do_sample_scipy.register
+@do_sample_scipy.register(GammaDistribution)
 def _(dist: GammaDistribution, size, seed):
     # https://stackoverflow.com/questions/42150965/how-to-plot-gamma-distribution-with-alpha-and-beta-parameters-in-python
     return scipy.stats.gamma.rvs(a=float(dist.k), scale=float(dist.theta), size=size, random_state=seed)
 
 
-@do_sample_scipy.register
+@do_sample_scipy.register(LogNormalDistribution)
 def _(dist: LogNormalDistribution, size, seed):
     # https://docs.scipy.org/doc/scipy/reference/generated/scipy.stats.lognorm.html
     return scipy.stats.lognorm.rvs(scale=float(exp(dist.mean)), s=float(dist.std), size=size, random_state=seed)
 
 
-@do_sample_scipy.register
+@do_sample_scipy.register(NormalDistribution)
 def _(dist: NormalDistribution, size, seed):
     return scipy.stats.norm.rvs(loc=float(dist.mean), scale=float(dist.std), size=size, random_state=seed)
 
 
-@do_sample_scipy.register
+@do_sample_scipy.register(ParetoDistribution)
 def _(dist: ParetoDistribution, size, seed):
     # https://stackoverflow.com/questions/42260519/defining-pareto-distribution-in-python-scipy
     return scipy.stats.pareto.rvs(b=float(dist.alpha), scale=float(dist.xm), size=size, random_state=seed)
 
 
-@do_sample_scipy.register
+@do_sample_scipy.register(StudentTDistribution)
 def _(dist: StudentTDistribution, size, seed):
     return scipy.stats.t.rvs(df=float(dist.nu), size=size, random_state=seed)
 
 
-@do_sample_scipy.register
+@do_sample_scipy.register(UniformDistribution)
 def _(dist: UniformDistribution, size, seed):
     # https://docs.scipy.org/doc/scipy/reference/generated/scipy.stats.uniform.html
     return scipy.stats.uniform.rvs(loc=float(dist.left), scale=float(dist.right - dist.left), size=size, random_state=seed)
 
 
-@do_sample_scipy.register
+@do_sample_scipy.register(BetaDistribution)
 def _(dist: BetaDistribution, size, seed):
     # same parametrisation
     return scipy.stats.beta.rvs(a=float(dist.alpha), b=float(dist.beta), size=size, random_state=seed)
 
 
-@do_sample_scipy.register
+@do_sample_scipy.register(CauchyDistribution)
 def _(dist: CauchyDistribution, size, seed):
     return scipy.stats.cauchy.rvs(loc=float(dist.x0), scale=float(dist.gamma), size=size, random_state=seed)
 
 
 # DRV:
 
-@do_sample_scipy.register
+@do_sample_scipy.register(DiscreteDistributionHandmade)
 def _(dist: DiscreteDistributionHandmade, size, seed):
     from scipy.stats import rv_discrete
     from sympy import lambdify
@@ -117,44 +117,44 @@ def _(dist: DiscreteDistributionHandmade, size, seed):
     return scipy_rv.rvs(size=size, random_state=seed)
 
 
-@do_sample_scipy.register
+@do_sample_scipy.register(GeometricDistribution)
 def _(dist: GeometricDistribution, size, seed):
     return scipy.stats.geom.rvs(p=float(dist.p), size=size, random_state=seed)
 
 
-@do_sample_scipy.register
+@do_sample_scipy.register(LogarithmicDistribution)
 def _(dist: LogarithmicDistribution, size, seed):
     return scipy.stats.logser.rvs(p=float(dist.p), size=size, random_state=seed)
 
 
-@do_sample_scipy.register
+@do_sample_scipy.register(NegativeBinomialDistribution)
 def _(dist: NegativeBinomialDistribution, size, seed):
     return scipy.stats.nbinom.rvs(n=float(dist.r), p=float(dist.p), size=size, random_state=seed)
 
 
-@do_sample_scipy.register
+@do_sample_scipy.register(PoissonDistribution)
 def _(dist: PoissonDistribution, size, seed):
     return scipy.stats.poisson.rvs(mu=float(dist.lamda), size=size, random_state=seed)
 
 
-@do_sample_scipy.register
+@do_sample_scipy.register(SkellamDistribution)
 def _(dist: SkellamDistribution, size, seed):
     return scipy.stats.skellam.rvs(mu1=float(dist.mu1), mu2=float(dist.mu2), size=size, random_state=seed)
 
 
-@do_sample_scipy.register
+@do_sample_scipy.register(YuleSimonDistribution)
 def _(dist: YuleSimonDistribution, size, seed):
     return scipy.stats.yulesimon.rvs(alpha=float(dist.rho), size=size, random_state=seed)
 
 
-@do_sample_scipy.register
+@do_sample_scipy.register(ZetaDistribution)
 def _(dist: ZetaDistribution, size, seed):
     return scipy.stats.zipf.rvs(a=float(dist.s), size=size, random_state=seed)
 
 
 # FRV:
 
-@do_sample_scipy.register
+@do_sample_scipy.register(SingleFiniteDistribution)
 def _(dist: SingleFiniteDistribution, size, seed):
     # scipy can handle with custom distributions
