+++ sympy/stats/sampling/tests/test_sample_continuous_rv.py
@@ -1,6 +1,7 @@
+from sympy import exp, Interval, oo, Symbol
 from sympy.external import import_module
 from sympy.stats import Beta, Chi, Normal, Gamma, Exponential, LogNormal, Pareto, ChiSquared, Uniform, sample, \
-    BetaPrime, Cauchy, GammaInverse, GaussianInverse, StudentT
+    BetaPrime, Cauchy, GammaInverse, GaussianInverse, StudentT, Weibull, density, ContinuousRV
 from sympy.testing.pytest import skip, ignore_warnings, raises
 
 
@@ -92,3 +93,91 @@ def test_sample_pymc3():
                     assert sam in X.pspace.domain.set
             raises(NotImplementedError,
                    lambda: next(sample(Chi("C", 1), library='pymc3')))
+
+
+def test_sampling_gamma_inverse():
+    scipy = import_module('scipy')
+    if not scipy:
+        skip('Scipy not installed. Abort tests for sampling of gamma inverse.')
+    X = GammaInverse("x", 1, 1)
+    with ignore_warnings(UserWarning): ### TODO: Restore tests once warnings are removed
+        assert next(sample(X)) in X.pspace.domain.set
+
+
+def test_lognormal_sampling():
+    # Right now, only density function and sampling works
+    scipy = import_module('scipy')
+    if not scipy:
+        skip('Scipy is not installed. Abort tests')
+    with ignore_warnings(UserWarning): ### TODO: Restore tests once warnings are removed
+        for i in range(3):
+            X = LogNormal('x', i, 1)
+            assert next(sample(X)) in X.pspace.domain.set
+
+    size = 5
+    with ignore_warnings(UserWarning): ### TODO: Restore tests once warnings are removed
+        samps = next(sample(X, size=size))
+        for samp in samps:
+            assert samp in X.pspace.domain.set
+
+
+def test_sampling_gaussian_inverse():
+    scipy = import_module('scipy')
+    if not scipy:
+        skip('Scipy not installed. Abort tests for sampling of Gaussian inverse.')
+    X = GaussianInverse("x", 1, 1)
+    with ignore_warnings(UserWarning): ### TODO: Restore tests once warnings are removed
+        assert next(sample(X, library='scipy')) in X.pspace.domain.set
+
+
+def test_prefab_sampling():
+    scipy = import_module('scipy')
+    if not scipy:
+        skip('Scipy is not installed. Abort tests')
+    N = Normal('X', 0, 1)
+    L = LogNormal('L', 0, 1)
+    E = Exponential('Ex', 1)
+    P = Pareto('P', 1, 3)
+    W = Weibull('W', 1, 1)
+    U = Uniform('U', 0, 1)
+    B = Beta('B', 2, 5)
+    G = Gamma('G', 1, 3)
+
+    variables = [N, L, E, P, W, U, B, G]
+    niter = 10
+    size = 5
+    with ignore_warnings(UserWarning): ### TODO: Restore tests once warnings are removed
+        for var in variables:
+            for _ in range(niter):
+                assert next(sample(var)) in var.pspace.domain.set
+                samps = next(sample(var, size=size))
+                for samp in samps:
+                    assert samp in var.pspace.domain.set
+
+
+def test_sample_continuous():
+    z = Symbol('z')
+    Z = ContinuousRV(z, exp(-z), set=Interval(0, oo))
+    assert density(Z)(-1) == 0
+
+    scipy = import_module('scipy')
+    if not scipy:
+        skip('Scipy is not installed. Abort tests')
+    with ignore_warnings(UserWarning): ### TODO: Restore tests once warnings are removed
+        assert next(sample(Z)) in Z.pspace.domain.set
+    sym, val = list(Z.pspace.sample().items())[0]
+    assert sym == Z and val in Interval(0, oo)
+
+    libraries = ['scipy', 'numpy', 'pymc3']
+    for lib in libraries:
+        try:
+            imported_lib = import_module(lib)
+            if imported_lib:
+                s0, s1, s2 = [], [], []
+                s0 = list(sample(Z, numsamples=10, library=lib, seed=0))
+                s1 = list(sample(Z, numsamples=10, library=lib, seed=0))
+                s2 = list(sample(Z, numsamples=10, library=lib, seed=1))
+                assert s0 == s1
+                assert s1 != s2
+        except NotImplementedError:
+            continue
